MOB TP MOVES (STAGE FINAL) : Scripting support for monster TP moves. Other fixes too. This is a large update so please check for bugs!
Details: Scripting support for monster TP moves has been added. New folder scripts/globals/mobskills to put them in. Calling function OnMobWeaponSkill. See the examples for information on physical, magical and enfeebling moves. There is a bug with setting HP to 0 for a monster which needs to be resolved. Also fixed an issue with a crash with magic. Added a new global for all mob TP moves (monstertpmoves.lua). Like I said, there may be teething problems with this update, I've added in new lua objects which I'm not sure if I've done correctly for the SVN, so be aware! Other than that, you'll see an error in console when mobs use TP moves now if the file isn't found. Hope you enjoy it >_<