RECAST UPDATE :: Fast Cast now works, reducing cast and recast times appropriately. In addition, Haste now reduces recast timers in addition to weapon delay.
There are caps in force for Fast Cast and Haste so you can't get 0 recast or anything. Speaking of 0 recast, I have had to nerf Chainspell a little -> retail does have a small period of delay between casts, which is now reflected as of this update. You will need to update traits.sql in order to get Fast Cast traits on RDM!