COUNTER :: Counter has been added, Counterstance has also been added, and various bug fixes. See details.
[Counter]
The counter trait and counter+ from equipment is used to determine your counter rate. However, there is an issue whereby a successfully countered attack WILL NOT DISPLAY ANY DAMAGE (hits for 0) even though it doesn't. The cause of this is under examination.

[Counterstance]
This has been added in the core with minor additions in LUA. It HAD to be done in the core due to glitchy behaviour of MOD_DEF when changing equipment. This is because Counterstance sets your DEF to VIT/2 and ignore DEF from other sources. You cannot use :setMod(VIT/2) else when it comes to remove the effect, you need to know how much to add back on (variable). You can't store how much to add back on because it may change over time as you equip/remove gear (which can result in 9999 DEF because of a loopback), so the only real method was to check in the core DEF() function, which is what Counterstance does.

[Fixes]
Linkpearl holders can no longer create a linkpearl by using /makelinkshell. A bug whereby Healing Breath when the DRG is not fighting a monster would crash the server has been fixed.