BUG FIXES :: Monster damage fixed and crafting bug fixed, see details.
Details: An issue whereby monsters would hit for extremely low damage when moderate amounts of DEF was equipped has been fixed. The cause has been an issue in calculating cRatio, which ended up being an integer calculation (so if you had enough DEF, the monsters cRatio would be say 0.98 which would resolve to 0). This should also technically apply to PCs so you should now get representative damage when hitting monsters as well. Also fixed the mentioned crafting bug when talking to an NPC when crafting. The core now simply sends a release packet and drops the npc trigger request from the client. Finally, decreased the probability of a monster >25% HP using a TP move >=100% from 45% to 40% per swing.