BCNM UPDATE (SEMI-FINAL) :: You can now script BCNMs using the new system, please see details to make the transition smooth!
THIS IS VERY MUCH IN AN ALPHA STAGE SO EXPECT BUGS.

[Pre-requisites to implementing BCNMs (DATABASE)]
You MUST have the database entries for the BCNMs you wish to script. Bluekirby is working on this. A brief explanation of the whole db end follows:

bcnm_info table:
- bcnmid = arbitrary number used to distinguish unique bcnms (with different monsters/rules/limitations)
- zoneid = the zone id for this bcnm
- name = the script name you want to call for the callbacks
- fastestName / fastestTime = null for now. Leave this alone.
- timelimit = The time limit in seconds
- levelcap = the level cap (currently NOT implemented)
- partysize = the pt size limit (1,3,6 are valid currently)
- lootDropId = the drop id for any treasure chests (not implemented currently)
- rules = a bitmask containing rules (see the comments in bcnm_info.sql)

bcnm_instance table:
bcnmid = the bcnmid which the monster belongs to
instancenumber = the instance number the monster belongs to (based on the SPAWN POSITION of the monster. Get this wrong and you may spawn Maat for instance #2 and send your players in instance #1! No good!)
monsterid = The monster ID to spawn.

[Scripting BCNMs]
An example script is given for Maat in Horlais Peak (handles entry/exit/spawning/etc). The file Burning_Circle.lua is largely omitted by the first few lines, it is mainly used for calling the global bcnm.lua file. This will be the same for most (all?) burning circle npcs. I have left the old code in for now, but once the transfer is complete, the npc files will just be one-liners which call the bcnm.lua global functions, much like Fields of Valor does for Field Manuals.

The bcnms/ folder has files which you named in the database. This is mainly used to give ending cutscenes if you win, but you can give key items etc. Most of the time this will be blank. If you win a BCNM, you must give the exit cutscene!

The global bcnm.lua file still contains the bulk of the work. For all TRADE BCNMs, look at itemid_bcnmid_map and add in entries which map the traded item id to the correct bcnm id. Do not include Maat fights (since you need to check job/level which is unique for BCNMs). Once you've done that, you need to find the correct bit to set for the event update. This is set using bcnmid_param_map, which maps the bcnmid from the database to the event parameter needed to display to the client. That SHOULD be all that needs to be done in bcnm.lua (just those 2 arrays at the top) and the rest just works. For non-traded BCNMs (Missions), you will need to edit EventTriggerBCNM to check if the mission is active, set the param, etc.

[Special BCNMs]
Some BCNMs have special winning requirements (e.g. steal from Maat, drop the monster to XX% HP). There is support for this but only in the core, under instanceutils::meetsWinningConditions or instanceutils::meetsLosingConditions depending on what special rule you want to implement.

[Limitations]
Records are not stored or displayed. If all 3 instances are full, the messages aren't completely correct and there are not thorough checks in place (potential for a crash). This can only be resolved when the correct update event parameters are found to tell the client that all instances are full and to NOT teleport them to the arena. 

[Main Features]
- Time limits enforced.
- Death rules enforced (remove after 3 min, though you can remove immediately by setting the bit in the rules field in bcnm_info.sql)
- Party handling (1, 3 or 6 people can join depending on what you specify in the database)
- Monster BCNM behaviour (they won't chase you out the BCNM)
- Instance handling (in allocation and transporting, it should be largely transparent to a scripter, in other words you shouldn't need to know (except when testing the monster positions))
- Extreme conditions handled (warping from a bcnm you're in, disconnecting from a bcnm you're in, etc)
 NOTE: If you d/c, you will be warped to the lobby. This is done in the CORE, and has only been done for Horlais Peak at current (I need to find the xyz for the other zones!)

[Summary to add new BCNMs]
- Add entries in database (bcnm_info and bcnm_instance)
- Edit the two arrays in bcnm.lua to specify the correct traded item and the correct update parameter based on the bcnmid you made via the database.
- Create and edit the script in bcnms/ with the winning cutscene, giving of key items, etc.
- TO DO NON-TRADED BCNMs: Edit EventTriggerBCNM to check for the mission, see if it's all valid then use startEvent with the correct event parameter. You will also need to use player:setVar("trade_bcnmid",id) to store the BCNM ID for onEventUpdate.

[Non-BCNM Related]
- Fixed a bug whereby you would crash if you needed a cutscene on zone.