TREASURE HUNTER :: Has been added, check the details for the calculations.
Treasure Hunter increases the odds of getting an item by granting an extra chance at getting an item. This provides a very stable non-linear increase in drops. Compare this to say a static multiplier (e.g. +50% relative drop per TH) which would mess up on higher % drops (e.g. a 66% chance with +50% relative drop would boost the drop rate to 66+33% = 99%, and any drops >66% would have over a 100% drop rate!!). Extra chances however do not do this, take for example:

10% drop rate. TH1 grants 1 "extra chance" of getting the item. You can either get the drop on the first go (10% chance) or you can get it on your 2nd go (0.9 * 10%) = 19% drop rate with TH1.

40% drop rate. TH2 grants 2 "extra chances":
0.4 + 0.6*0.4 + 0.6*0.6*0.4 = 0.784 = 78.4% chance

90% drop rate. TH2 grants 2 "extra chances":
0.9 + 0.1*0.9 + 0.1*0.1*0.9 =  99.9% chance

Of course, this is only the expected chance, it totally possible to go several mobs without getting a 99.9% drop. Only the highest TH in the party is taken into account. The freaky "additional effect TH" has not been implemented. Furthermore, traits on non-THF jobs may say Treasure Hunter III but this is incorrect. Lv75+ traits are displayed incorrectly to the player in some cases.