WEAPONSKILLS :: "Critical hit varies with TP" Weaponskills have been enhanced.
Or rather, implemented. It seems that the values/boolean were not set to true for the vast majority of critical hit WSes. This meant their damage was well below what you would expect. The weaponskills involved are:
- Vorpal blade
- Raging rush
- Evisceration 
- Skewer 

Furthermore, critical hit WSes which only have a single hit (not multi hit) have had their odds of scoring a critical hit boosted. This includes:
- Blade: Rin
- Keen Edge
- Vorpal Scythe
- Vorpal Thrust

Finally, checks are in place for lower bounds of pdif, so you should not have incorrect damage displayed on WSes now.