SCRIPTING: - Lure of the Wildcat (Jeuno) now implemented.  See note. - Some assorted refactoring along the way which never breaks anything, ever. - The Merit unlock cutscene should now be the correct one.
NOTE:
This version of Lure of the Wildcat uses the ENABLE_TOAU variable in settings.lua, so server owners can easily deny access to ToAU.  Make sure you enable it and restart your server before testing the quest.
I was unable to find the cutscene for Aht Urghan mission 1, and I wasn't throwing all this work away.  Mission progression is not yet required, and can easily be re-enabled later.