CORE: Monsters will now cast AoE magic. (Tier I-V)
Due to the fact that RDMs don't get AoE magic, it is therefore restricted to BLMs only, and WHMs with Banishga. There will need to be some logic fudging if you set fire.aoe = 1 and they are RDM, perhaps by going off the BLM tables rather than the RDM tables.