CORE: reimplemented HP/TP latent effects (excluding parade gorget, this time hopefully without crashing every server dozens of times) full list: http://kile.dyndns.info:8765/paste/index.php?show=3975
for the curious: defining a virtual function adds 4 bytes for a virtual function table pointer at the start of a class.  the issue is that CLuaBaseEntity is based off CBaseEntity: which didn't have a virtual function table pointer.  So any time a CBattleEntity or its subclass were pushed into lua, it wasn't expecting the 4 extra bytes.  So when the core got the entity back (via new CLuaBaseEntity), the BaseEntity was corrupt because it was offset by 4 bytes.  Solved this by making a function in CBaseEntity virtual, to give it the virtual table pointer.  Not really ideal (since it increases size for no reason) but there wasn't really a way around it without refactoring half the lua bindings