SCRIPTS: Adds TIMEZONE_OFFSET to settings.lua for locale management. - Two new functions have been added to common.lua named getMidnight() and getConquestTally(), both using TIMEZONE_OFFSET to provide Japanese Midnight by default.  These can be used anywhere in LUA.
Usage:
getMidnight() will return the end of the current real day.  player:setVar("QUEST_VARIABLE_WAIT",getMidnight()), if os.time() < player:getVar("WAIT_ONE_DAY") then, give player cutscene to wait, else, progress quest.  Ceraulian in Port San d'Oria has been updated to use this method as an example.
getConquestTally() will return the end of the current conquest period (24:00 Saturday), and can be used in the same way.

Updating old quests:
It should simply be a matter of making sure that os.time() is either greater than the variable, or that the variable is less than os.time().  Since they used to be stored in day-of-the-year format (max 366?), this shouldn't be an issue.  Worst case scenario, anyone with the quest currently active won't have to actually wait until midnight.