CONFIG: Removed server message from being stored inside of map_darkstar.conf CONFIG: Added new server message files for English and French. (server_message.conf and server_message_fr.conf) CORE: Recoded how server messages are handled to properly allow the client to request longer messages.
Note: There is still an issue with /servmes crashing the client with random server messages. I cannot pinpoint this issue at the moment but I am certain it is not the packets but the client. (Perhaps it is an issue with us not emulating all of POL properly in the loader.)