Add ACP mission: Gatherer of Light (I)
Also included:
Partially scripted Squintrox Dryeyes and Tenshodo Treasure Coffer. What is already scripted can be used as a sort of template for the the rest.

I advise keeping the add-on disabled in settings.lua until augments are working in core.

The augmented rewards will not give stats until augments are in the core, but items acquired through this script WILL work properly when that happens and will display (what would be) the correct stats on the item in the meantime. I tripple checked all Augment IDs and Augment values.

A note regarding key item drops:
When I was farming in retail the goblins that have pets would not give up the key item until both pet and master were killed. I'd often get the Ki from the goblins bee, rather than the goblin, but ONLY if the goblin was already dead. I presently just have the goblin dropping the keyitem. Other than that the scripting behaves as retail as can be. I had to guess at the drop rate and it may be a little high or low compared to retail.