Implement Auspice Enspell & Accuracy Bonus w/ Afflatus Misery
Enspell Bonus:
If Afflatus Misery & Auspice after active together, then auspice acts as a light enspell. It is calculated like at tier II enspell and must check that it applies only on hitCount 0

Accuracy Bonus:
Each time the player misses an attack while the combination of Misery & Auspice is active, the system is supposed to boost their accuracy by 10. This effect wears off as soon as eith Auspice OR Misery wears off.

Since I am still new, I poked around looking for ways to do this, so this could be noobish. The way I found that works is to track the accumlating bonus as a char_var while adding the bonus to MOD_ACC. The char_var is then used to subtract the bonus that was gained from the combination from MOD_ACC when either effect wears off. If this is not an appropriate way to do this, please let me know. I am always happy to learn.